+++ internal/libs/confix/plan.go
@@ -1,4 +1,4 @@
-package main
+package confix
 
 import (
 	"context"
