+++ internal/libs/confix/testdata/README.md
@@ -41,12 +41,12 @@ The files named `diff-XX-YY.txt` were generated by using the `condiff` tool on
 the config samples for versions v0.XX and v0.YY:
 
 ```shell
-go run ./scripts/confix/condiff -desnake vXX-config vYY-config.toml > diff-XX-YY.txt
+go run ./scripts/condiff -desnake vXX-config vYY-config.toml > diff-XX-YY.txt
 ```
 
 The `baseline.txt` was computed in the same way, but using an empty starting
 file so that we capture all the settings in the target:
 
 ```shell
-go run ./scripts/confix/condiff -desnake /dev/null v26-config.toml > baseline.txt
+go run ./scripts/condiff -desnake /dev/null v26-config.toml > baseline.txt
 ```
